feat(dto): update data types and validation in UpdateUserInput DTO

+++ internal/dto/v1/user_dto.go
@@ -29,10 +29,10 @@ type CreateUserInput struct {
 
 type UpdateUserInput struct {
 	Name     *string `json:"name" binding:"omitempty"`
-	Age      *int    `json:"age" binding:"omitempty,gt=18"`
+	Age      *int32  `json:"age" binding:"omitempty,gt=18"`
 	Password *string `json:"password" binding:"omitempty,min=8,password_strong"`
-	Status   *int    `json:"status" binding:"omitempty,oneof=1 2 3"`
-	Level    *int    `json:"level" binding:"omitempty,oneof=1 2 3"`
+	Status   *int32  `json:"status" binding:"omitempty,oneof=1 2 3"`
+	Level    *int32  `json:"level" binding:"omitempty,oneof=1 2 3"`
 }
 
 // Request
@@ -52,8 +52,8 @@ func (input *UpdateUserInput) MapUpdateInputToModel(userUuid uuid.UUID) sqlc.Upd
 		UserPassword: input.Password,
 		UserFullname: input.Name,
 		UserAge:      utils.ConvertToInt32Pointer(input.Age),
-		UserStatus:   int32(input.Status),
-		UserLevel:    int32(input.Level),
+		UserStatus:   input.Status,
+		UserLevel:    input.Level,
 	}
 }
